refactor: Order package 및 의존성 분리
- OrderValidator 로 의존성 약화
