+++ src/main/java/kitchenpos/domain/OrderTables.java
@@ -1,6 +1,9 @@
 package kitchenpos.domain;
 
+import kitchenpos.table.domain.OrderTable;
+
 import javax.persistence.Embeddable;
+import javax.persistence.JoinColumn;
 import javax.persistence.OneToMany;
 import java.util.ArrayList;
 import java.util.List;
@@ -8,7 +11,8 @@ import java.util.List;
 @Embeddable
 public class OrderTables {
 
-    @OneToMany(mappedBy = "tableGroup")
+    @OneToMany
+    @JoinColumn(name = "table_group_id")
     private List<OrderTable> orderTables = new ArrayList<>();
 
     public List<OrderTable> getOrderTables() {
