+++ src/main/java/kitchenpos/order/dto/OrderLineItemResponseDto.java
@@ -1,6 +1,6 @@
-package kitchenpos.dto;
+package kitchenpos.order.dto;
 
-import kitchenpos.domain.OrderLineItem;
+import kitchenpos.order.domain.OrderLineItem;
 
 public class OrderLineItemResponseDto {
     private Long seq;
@@ -9,7 +9,7 @@ public class OrderLineItemResponseDto {
 
     public OrderLineItemResponseDto(OrderLineItem orderLineItem) {
         this.seq = orderLineItem.getSeq();
-        this.menuId = orderLineItem.getMenu().getId();
+        this.menuId = orderLineItem.getMenuId();
         this.quantity = orderLineItem.getQuantity();
     }
