+++ src/main/java/kitchenpos/order/dto/OrderRequestDto.java
@@ -1,6 +1,10 @@
-package kitchenpos.dto;
+package kitchenpos.order.dto;
+
+import kitchenpos.order.domain.Order;
+import kitchenpos.order.domain.OrderLineItem;
 
 import java.util.List;
+import java.util.stream.Collectors;
 
 public class OrderRequestDto {
     private Long orderTableId;
@@ -20,5 +24,12 @@ public class OrderRequestDto {
 
     public List<OrderLineItemRequestDto> getOrderLineItems() {
         return orderLineItems;
+    }
+
+    public Order toEntity() {
+        List<OrderLineItem> orderLineItems = this.orderLineItems.stream()
+                .map(it -> new OrderLineItem(it.getMenuId(), it.getQuantity()))
+                .collect(Collectors.toList());
+        return new Order(orderTableId, orderLineItems);
     }
 }
