+++ src/main/java/kitchenpos/table/domain/OrderTable.java
@@ -2,10 +2,7 @@ package kitchenpos.table.domain;
 
 import kitchenpos.domain.TableGroup;
 
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
+import javax.persistence.*;
 
 @Entity
 public class OrderTable {
@@ -17,6 +14,7 @@ public class OrderTable {
     private Long id;
 
 
+    @Column(name = "table_group_id")
     private Long tableGroupId;
     private int numberOfGuests;
     private boolean empty;
