+++ src/test/java/kitchenpos/fixture/OrderLineItemFixture.java
@@ -1,9 +1,8 @@
 package kitchenpos.fixture;
 
-import kitchenpos.domain.OrderLineItem;
-import kitchenpos.dto.OrderLineItemRequestDto;
-import kitchenpos.dto.OrderLineItemResponseDto;
-import kitchenpos.menu.domain.Menu;
+import kitchenpos.order.domain.OrderLineItem;
+import kitchenpos.order.dto.OrderLineItemRequestDto;
+import kitchenpos.order.dto.OrderLineItemResponseDto;
 
 public class OrderLineItemFixture {
 
@@ -11,8 +10,8 @@ public class OrderLineItemFixture {
         return new OrderLineItemRequestDto(menuId, quantity);
     }
 
-    public static OrderLineItem 주문항목_데이터_생성(Long seq, Menu menu, long quantity) {
-        return new OrderLineItem(seq, menu, quantity);
+    public static OrderLineItem 주문항목_데이터_생성(Long seq, Long menuId, long quantity) {
+        return new OrderLineItem(seq, menuId, quantity);
     }
 
     public static OrderLineItemResponseDto 주문항목_응답_데이터_생성(Long seq, Long menuId, int quantity) {
