+++ src/test/java/kitchenpos/fixture/OrderTableFixture.java
@@ -1,10 +1,8 @@
 package kitchenpos.fixture;
 
-import kitchenpos.domain.OrderTable;
-import kitchenpos.domain.TableGroup;
-import kitchenpos.dto.OrderTableRequestDto;
-import kitchenpos.dto.OrderTableResponseDto;
-import kitchenpos.dto.TableGroupResponseDto;
+import kitchenpos.table.domain.OrderTable;
+import kitchenpos.table.dto.OrderTableRequestDto;
+import kitchenpos.table.dto.OrderTableResponseDto;
 
 public class OrderTableFixture {
 
@@ -12,14 +10,12 @@ public class OrderTableFixture {
         return new OrderTableRequestDto(numberOfGuests, false);
     }
 
-    public static OrderTable 주문테이블_데이터_생성(Long id, TableGroup tableGroup, int numberOfGuests, boolean isEmpty) {
-        return new OrderTable(id, tableGroup, numberOfGuests, isEmpty);
+    public static OrderTable 주문테이블_데이터_생성(Long id, Long tableGroupId, int numberOfGuests, boolean isEmpty) {
+        return new OrderTable(id, tableGroupId, numberOfGuests, isEmpty);
     }
 
-    public static OrderTableResponseDto 주문테이블_응답_데이터_생성(Long id, int numberOfGuests, boolean isEmpty,
-        TableGroupResponseDto tableGroup) {
-
-        return new OrderTableResponseDto(id, numberOfGuests, isEmpty ,tableGroup);
+    public static OrderTableResponseDto 주문테이블_응답_데이터_생성(Long id, int numberOfGuests, boolean isEmpty, Long tableGroupId) {
+        return new OrderTableResponseDto(id, numberOfGuests, isEmpty ,tableGroupId);
     }
 
 }
