+++ src/test/java/kitchenpos/repository/OrderRepositoryTest.java
@@ -1,13 +1,15 @@
 package kitchenpos.repository;
 
-import static org.junit.jupiter.api.Assertions.assertEquals;
-
-import java.util.Arrays;
 import kitchenpos.domain.OrderStatus;
+import kitchenpos.order.repository.OrderRepository;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
+
+import java.util.Arrays;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
 
 @DataJpaTest
 class OrderRepositoryTest {
