+++ src/test/java/kitchenpos/repository/OrderTableRepositoryTest.java
@@ -1,6 +1,7 @@
 package kitchenpos.repository;
 
-import kitchenpos.domain.OrderTable;
+import kitchenpos.table.domain.OrderTable;
+import kitchenpos.table.repository.OrderTableRepository;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
